make mhp examples
