+++ examples/mhp/CMakeLists.txt
@@ -1,0 +1,23 @@
+# SPDX-FileCopyrightText: Intel Corporation
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+add_executable(vector-add vector-add.cpp)
+target_link_libraries(vector-add DR::mpi)
+add_mpi_test(vector-add vector-add 2)
+
+add_executable(stencil-1d stencil-1d.cpp)
+target_link_libraries(
+  stencil-1d
+  cxxopts
+  DR::mpi)
+add_mpi_test(stencil-1d stencil-1d 2)
+
+if(OpenMP_FOUND)
+  add_executable(vector-add-ref vector-add-ref.cpp)
+  target_link_libraries(vector-add-ref PRIVATE MPI::MPI_CXX OpenMP::OpenMP_CXX fmt::fmt)
+  add_mpi_test(vector-add-ref vector-add-ref 2)
+endif()
+
+add_executable(transpose-ref transpose-ref.cpp)
+target_link_libraries(transpose-ref PRIVATE MPI::MPI_CXX MKL::MKL)
