+++ examples/mhp/transpose-ref.cpp
