+++ examples/mhp/vector-add-ref.cpp
