+++ examples/shp/take_example.cpp
