+++ examples/shp/test_range.cpp
