+++ examples/shp/vector_example.cpp
